requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# LangChain and LangGraph dependencies - compatible versions
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langgraph==0.2.0
langchain-google-genai==1.0.10
langchain-google-vertexai==1.0.10

# Vector database and embeddings
chromadb==0.4.24
sentence-transformers==2.7.0

# Google AI dependencies
google-generativeai==0.7.2
google-cloud-aiplatform==1.58.0

# Hugging Face dependencies
transformers==4.42.0
torch>=2.2.0
accelerate==0.32.0

# Data processing and utilities
pandas==2.2.0
numpy==1.26.0
requests==2.32.0
aiohttp==3.9.5

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.8.0
pydantic-settings==2.4.0

# Logging and monitoring
structlog==23.2.0

# Development dependencies
pytest==8.3.0
pytest-asyncio==0.23.8
black==24.8.0
flake8==7.1.0

# Web Search
duckduckgo-search-8.1.1
lxml-6.0.1
primp-0.15.0

